+++ packages/security/src/globals/utils/events.js
@@ -1,5 +1,5 @@
 /**
- * Copyright IBM Corp. 2016, 2018
+ * Copyright IBM Corp. 2016, 2021
  *
  * This source code is licensed under the Apache-2.0 license found in the
  * LICENSE file in the root directory of this source tree.
@@ -21,11 +21,9 @@ const composeEventHandlers =
       if (event.defaultPrevented) {
         break;
       }
-      /* eslint-disable security/detect-object-injection */
       if (typeof fns[i] === 'function') {
         fns[i](event, ...args);
       }
-      /* eslint-enable security/detect-object-injection */
     }
   };